requirements.txt
flask
flask-login
flask-sqlalchemy
flask-migrate
python-dotenv
werkzeug
